requirements.txt
asgiref==3.3.4
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
Django==3.2.3
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
idna==2.10
more-itertools==8.8.0
packaging==20.9
Pillow==7.0.0
pluggy==0.13.1
py==1.10.0
PyJWT==2.1.0
pyparsing==2.4.7
pytest==5.4.1
pytest-django==3.9.0
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
six==1.14.0
sqlparse==0.4.1
urllib3==1.26.5
wcwidth==0.2.5
pytils==0.3
gunicorn==20.1.0
psycopg2-binary==2.8.6
celery==5.1.2
flake8==3.9.2
django-redis==5.0.0
flower==1.0.0
django-celery-beat==2.2.1